runtime: Use relative rate numerator and denominator integers when appropriate
In several places in the gnuradio-runtime, use the d_rr_interpolation and
d_rr_decimation integer values instead of the the floating point
d_relative_rate value.  This should make tag propagation of tags with
large offsets precise for integer relative rates that cannot
be expressed with a terminating fractional part in a finite precision double
(e.g. 1/5 is non-terminating when expressed in binary.)